+++ core/src/main/java/uk/ac/bbsrc/tgac/miso/core/data/impl/view/qc/RunPartitionQcNode.java
@@ -64,9 +64,6 @@ public class RunPartitionQcNode implements QcNode {
 
   private static final long serialVersionUID = 1L;
 
-  private static final String ENTITY_TYPE = "RunPartition";
-  private static final String TYPE_LABEL = "Partition";
-
   @Id
   @ManyToOne
   @JoinColumn(name = "runId")
@@ -138,13 +135,13 @@ public class RunPartitionQcNode implements QcNode {
   }
 
   @Override
-  public String getEntityType() {
-    return ENTITY_TYPE;
+  public QcNodeType getEntityType() {
+    return QcNodeType.RUN_PARTITION;
   }
 
   @Override
   public String getTypeLabel() {
-    return TYPE_LABEL;
+    return QcNodeType.RUN_PARTITION.getLabel();
   }
 
   @Override